+++ internal/protoplugin/protoplugin.go
@@ -45,6 +45,7 @@ import (
 
 	"github.com/gogo/protobuf/proto"
 	"github.com/gogo/protobuf/protoc-gen-gogo/descriptor"
+	gogogen "github.com/gogo/protobuf/protoc-gen-gogo/generator"
 	"github.com/gogo/protobuf/protoc-gen-gogo/plugin"
 )
 
@@ -199,7 +200,11 @@ func (m *Message) FQMN() string {
 func (m *Message) GoType(currentPackage string) string {
 	var components []string
 	components = append(components, m.Outers...)
-	components = append(components, m.GetName())
+	// gogo_protobuf uses CamelCaseSlice which internally uses CamelCase for the GoType name conversion.
+	// see: https://github.com/gogo/protobuf/blob/v1.3.1/protoc-gen-gogo/generator/generator.go#L1810
+	// Added gogogen.CamelCase({message_name}) to keep GoTypes consistent between gogo_protobuf and yarpc generated
+	// protobuf files.
+	components = append(components, gogogen.CamelCase(m.GetName()))
 
 	name := strings.Join(components, "_")
 	if m.File.GoPackage.Path == currentPackage {