+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.55.1] - 2021-07-14
+### Fixed
+- peer/direct: peer connections were closed even if they were still in use.
+- configuration: clarify error message for the special case of attempting to use a peer list
+  updater when none have been registered.
+- protoplugin: generated golang code from the yarpc plugin follows same GoType conventions as gogo/protobuf.  
+
 ## [1.55.0] - 2021-07-06
 - Downgrade github.com/apache/thrift to the previously-compatible version (0.10)
 
@@ -1371,6 +1378,7 @@ This release requires regeneration of ThriftRW code.
 
 - Initial release.
 
+[1.55.1]: https://github.com/yarpc/yarpc-go/compare/v1.55.0...v1.55.1
 [1.55.0]: https://github.com/yarpc/yarpc-go/compare/v1.54.2...v1.55.0
 [1.54.2]: https://github.com/yarpc/yarpc-go/compare/v1.54.1...v1.54.2
 [1.54.1]: https://github.com/yarpc/yarpc-go/compare/v1.54.0...v1.54.1